refactor: domain 패키지 이름을 entity 로 변경
