+++ src/main/java/roomescape/reservation/repository/ReservationRepository.java
@@ -3,7 +3,7 @@ package roomescape.reservation.repository;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Optional;
-import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.entity.Reservation;
 
 public interface ReservationRepository {
