+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -5,9 +5,9 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.reservation.repository.ReservationRepository;
-import roomescape.theme.domain.Theme;
 import roomescape.theme.dto.request.ThemeRequestDTO;
 import roomescape.theme.dto.response.ThemeResponseDTO;
+import roomescape.theme.entity.Theme;
 import roomescape.theme.exception.AlreadyReservedThemeException;
 import roomescape.theme.exception.DuplicatedThemeException;
 import roomescape.theme.exception.NoSuchThemeException;
