+++ src/test/java/roomescape/reservation/repository/ReservationJdbcTemplateRepositoryTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.entity.Reservation;
 import roomescape.reservation.exception.NoSuchReservationException;
 
 @JdbcTest
