+++ src/test/java/roomescape/theme/service/ThemeServiceTest.java
@@ -11,10 +11,10 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.entity.Reservation;
 import roomescape.reservation.repository.ReservationRepository;
-import roomescape.theme.domain.Theme;
 import roomescape.theme.dto.request.ThemeRequestDTO;
+import roomescape.theme.entity.Theme;
 import roomescape.theme.exception.AlreadyReservedThemeException;
 import roomescape.theme.exception.DuplicatedThemeException;
 import roomescape.theme.exception.NoSuchThemeException;
